[ipSDK] BuildHelper  - 대화상자 로직 개발.
git-svn-id: https://192.168.6.225:8443/svn/CAD/Tag20120904/Utils/BuildHelper@42294 d8ace140-e149-8147-8d3d-8bcd6e031801

+++ BuildHelper/Job.cpp
@@ -112,17 +112,17 @@ UINT Job::DoModal()
 	case JobBase::EN_JOB_TYPE_FILECOPY:
 	{
 		JobFileCopyDlg dlg2((FileCopy*)GetImpl());
-		return dlg2.DoModal();
+		return (UINT)dlg2.DoModal();
 	} break;
 	case JobBase::EN_JOB_TYPE_FILEEXECUTE:
 	{
 		JobFileExecuteDlg dlg2((FileExecute*)GetImpl());
-		return dlg2.DoModal();
+		return (UINT)dlg2.DoModal();
 	} break;
 	case JobBase::EN_JOB_TYPE_FILEBATCH:
 	{
 		JobBatchDlg dlg2((FileBatch*)GetImpl());
-		return dlg2.DoModal();
+		return (UINT)dlg2.DoModal();
 	} break;
 	default:
 		ASSERT(0);
